+++ core/src/pipeline/mod.rs
@@ -4,4 +4,4 @@ pub mod segment_processing;
 pub mod segmentation_and_ocr;
 pub mod structured_extraction;
 pub mod update_metadata;
-pub mod upsert_output;
+pub mod upload_objects;
