+++ core/src/routes/task.rs
@@ -2,8 +2,8 @@ use crate::models::chunkr::auth::UserInfo;
 use crate::models::chunkr::task::Configuration;
 use crate::models::chunkr::task::TaskResponse;
 use crate::models::chunkr::upload::UploadForm;
-use crate::utils::server::create_task::create_task;
-use crate::utils::server::get_task::get_task;
+use crate::utils::routes::create_task::create_task;
+use crate::utils::routes::get_task::get_task;
 use actix_multipart::form::MultipartForm;
 use actix_web::{web, Error, HttpRequest, HttpResponse};
 use uuid::Uuid;
