+++ core/src/utils/mod.rs
@@ -2,8 +2,8 @@ pub mod clients;
 pub mod jobs;
 pub mod rate_limit;
 pub mod retry;
+pub mod routes;
 pub mod rrq;
-pub mod server;
 pub mod services;
 pub mod storage;
 pub mod stripe;
