+++ core/src/utils/routes/get_user.rs
